Revert "Vis årsaker  tilgangsfeil (#8824)"
This reverts commit 5e5519309b702334bb6594fe6c2aecc60fddb62d.
