+++ packages/v2/backend/src/shared/errorhandling/ÅrsakIkkeTilgang.ts
@@ -1,52 +1,15 @@
-import { k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT1 } from '@k9-sak-web/backend/k9sak/generated/types.js';
-import { k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT2 } from '@k9-sak-web/backend/ungsak/generated/types.js';
-import { k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT3 } from '@k9-sak-web/backend/k9klage/generated/types.js';
+import {k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT1} from '@k9-sak-web/backend/k9sak/generated/types.js';
+import {k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT2} from '@k9-sak-web/backend/ungsak/generated/types.js';
+import {k9_felles_sikkerhet_abac_ÅrsakIkkeTilgang as ÅIT3} from '@k9-sak-web/backend/k9klage/generated/types.js';
 
 import { safeConstCombine } from '@k9-sak-web/backend/typecheck/safeConstCombine.js';
 
-export const ÅrsakIkkeTilgang = safeConstCombine(ÅIT1, ÅIT2, ÅIT3);
+export const ÅrsakIkkeTilgang = safeConstCombine(ÅIT1, ÅIT2, ÅIT3)
 
-export type ÅrsakIkkeTilgang = ÅIT1 | ÅIT2 | ÅIT3;
+export type ÅrsakIkkeTilgang = ÅIT1 | ÅIT2 | ÅIT3
 
 export const isÅrsakIkkeTilgang = (maybe: unknown): maybe is ÅrsakIkkeTilgang =>
-  typeof maybe === 'string' && Object.values(ÅrsakIkkeTilgang).some(v => v === maybe);
+  typeof maybe === 'string' && Object.values(ÅrsakIkkeTilgang).some(v => v === maybe)
 
 export const isÅrsakIkkeTilgangArray = (maybe: unknown): maybe is ReadonlyArray<ÅrsakIkkeTilgang> =>
-  maybe instanceof Array && maybe.every(v => isÅrsakIkkeTilgang(v));
-
-// Visningstekst for kvar årsak til at brukar ikkje har tilgang.
-const årsakIkkeTilgangTekst: Record<ÅrsakIkkeTilgang, string> = {
-  HAR_IKKE_TILGANG_TIL_KODE6_PERSON: 'Du mangler tilgang til personer med strengt fortrolig adresse (kode 6)',
-  HAR_IKKE_TILGANG_TIL_KODE7_PERSON: 'Du mangler tilgang til personer med fortrolig adresse (kode 7)',
-  HAR_IKKE_TILGANG_TIL_EGEN_ANSATT: 'Du mangler tilgang til saker som gjelder Nav-ansatte',
-  HAR_IKKE_TILGANG_TIL_HISTORISK_SAK: 'Du mangler tilgang til historiske saker',
-  HAR_IKKE_TILGANG_TIL_APPLIKASJONEN: 'Du er ikke tildelt en rolle som gir tilgang til k9-sak',
-  HAR_IKKE_TILGANG_TIL_TJENESTE_FOR_BORGER: 'Tjenesten er kun tilgjengelig for borgere, ikke Nav-ansatte',
-  HAR_IKKE_TILGANG_TIL_TJENESTE_FOR_DRIFT: 'Du mangler driftsrettigheter',
-  HAR_IKKE_TILGANG_TIL_PIP_TJENESTE: 'Du har ikke tilgang til PIP-tjenesten',
-  HAR_IKKE_TILGANG_ANNEN_GRUNN: 'Tilgang avslått av annen grunn',
-  TEKNISK_FEIL: 'Teknisk feil ved tilgangskontroll',
-};
-
-// Årsakene vi ønsker å vise til brukar. Andre årsaker filtrerast bort.
-const årsakerViØnskerÅVise: ÅrsakIkkeTilgang[] = [
-  'HAR_IKKE_TILGANG_TIL_KODE6_PERSON',
-  'HAR_IKKE_TILGANG_TIL_KODE7_PERSON',
-  'HAR_IKKE_TILGANG_TIL_EGEN_ANSATT',
-  'HAR_IKKE_TILGANG_TIL_HISTORISK_SAK',
-  'HAR_IKKE_TILGANG_TIL_APPLIKASJONEN',
-  'TEKNISK_FEIL',
-];
-
-// Filtrer årsaker til dei vi ønsker å vise, og fjern duplikater.
-const resolveÅrsakerIkkeTilgang = (ikkeTilgangÅrsaker?: ReadonlyArray<ÅrsakIkkeTilgang>): ÅrsakIkkeTilgang[] => {
-  const filtrerteÅrsaker = ikkeTilgangÅrsaker
-    ? ikkeTilgangÅrsaker.filter(årsak => årsakerViØnskerÅVise.includes(årsak))
-    : [];
-
-  return Array.from(new Set(filtrerteÅrsaker));
-};
-
-// Utled visningstekstane for årsakene vi ønsker å vise til brukar.
-export const resolveÅrsakIkkeTilgangTekster = (ikkeTilgangÅrsaker?: ReadonlyArray<ÅrsakIkkeTilgang>): string[] =>
-  resolveÅrsakerIkkeTilgang(ikkeTilgangÅrsaker).map(årsak => årsakIkkeTilgangTekst[årsak] ?? årsak);
+  maybe instanceof Array && maybe.every(v => isÅrsakIkkeTilgang(v))
