+++ packages/v2/gui/src/app/errorhandling/pages/ForbiddenPage.stories.tsx
@@ -25,7 +25,7 @@ export const FeilmeldingIkkeTilgangMedÅrsaker: Story = {
   play: async ({ canvas }) => {
     await expect(canvas.getByRole('heading', { name: 'Du har ikke tilgang til denne saken' })).toBeInTheDocument();
     await expect(
-      canvas.getByText('Du mangler tilgang til personer med strengt fortrolig adresse (kode 6)'),
+      canvas.getByText('Du mangler tilgang til saker med strengt fortrolig adresse (kode 6)'),
     ).toBeInTheDocument();
     await expect(canvas.getByText('Du mangler tilgang til historiske saker')).toBeInTheDocument();
   },
