+++ packages/v2/gui/src/app/errorhandling/ui/resolveAxiosErrorView.tsx
@@ -1,5 +1,5 @@
 import { AxiosError } from 'axios';
-import { BodyLong, List } from '@navikt/ds-react';
+import { BodyLong } from '@navikt/ds-react';
 import { EnterIcon } from '@navikt/aksel-icons';
 import { resolveLoginURL, withRedirectToCurrentLocation } from '@k9-sak-web/backend/shared/auth/resolveLoginURL.js';
 import { formatDate, timeFormat } from '@k9-sak-web/gui/utils/formatters.js';
@@ -8,7 +8,6 @@ import { reloadAction, reloadActionWithFormResetWarning, restartAction } from '.
 import { BlobResponseAxiosError } from '../legacycompat/BlobResponseAxiosError.js';
 import {
   isÅrsakIkkeTilgangArray,
-  resolveÅrsakIkkeTilgangTekster,
   type ÅrsakIkkeTilgang,
 } from '@k9-sak-web/backend/shared/errorhandling/ÅrsakIkkeTilgang.js';
 
@@ -100,20 +99,10 @@ export const resolveAxiosErrorView = (error: AxiosError): ErrorViewProps => {
 
   // 403 — Tilgang nektet.
   if (status === 403) {
-    const årsakstekster = resolveÅrsakIkkeTilgangTekster(resolveAxiosErrorÅrsakIkkeTilgang(error));
     return {
       error,
       title: 'Ikke tilgang',
-      errorInfo:
-        årsakstekster.length > 0 ? (
-          <List>
-            {årsakstekster.map(tekst => (
-              <List.Item key={tekst}>{tekst}</List.Item>
-            ))}
-          </List>
-        ) : (
-          <BodyLong>Du har ikke tilgang til å gjøre denne handlingen eller se denne informasjonen.</BodyLong>
-        ),
+      errorInfo: <BodyLong> Du har ikke tilgang til å gjøre denne handlingen eller se denne informasjonen. </BodyLong>,
       fixAction: {
         ...restartAction,
         info: (
